alt_las_gpu_test/requirements.txt
flask==2.0.1
werkzeug==2.0.1
pynvml==11.0.0
numpy==1.21.0
psutil==5.8.0
requests==2.26.0
pyyaml==6.0
matplotlib==3.4.3
